Realized basic refactoring for vnc talk variant of firebase plugin - add utils, separated model in other folder, refactored  dispay messages process, optimised saving data in files and in shared prefs
