+++ src/android/FirebasePlugin.java
@@ -1,38 +1,40 @@
 package org.apache.cordova.firebase;
 
 import android.app.NotificationManager;
-import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.support.v4.app.NotificationManagerCompat;
 import android.util.Base64;
 import android.util.Log;
 
 import com.crashlytics.android.Crashlytics;
-import io.fabric.sdk.android.Fabric;
-
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.FirebaseApp;
+import com.google.firebase.FirebaseException;
+import com.google.firebase.FirebaseTooManyRequestsException;
 import com.google.firebase.analytics.FirebaseAnalytics;
+import com.google.firebase.auth.FirebaseAuthInvalidCredentialsException;
+import com.google.firebase.auth.PhoneAuthCredential;
+import com.google.firebase.auth.PhoneAuthProvider;
 import com.google.firebase.iid.FirebaseInstanceId;
 import com.google.firebase.messaging.FirebaseMessaging;
+import com.google.firebase.perf.FirebasePerformance;
+import com.google.firebase.perf.metrics.Trace;
 import com.google.firebase.remoteconfig.FirebaseRemoteConfig;
 import com.google.firebase.remoteconfig.FirebaseRemoteConfigInfo;
 import com.google.firebase.remoteconfig.FirebaseRemoteConfigSettings;
 import com.google.firebase.remoteconfig.FirebaseRemoteConfigValue;
-import com.google.firebase.perf.FirebasePerformance;
-import com.google.firebase.perf.metrics.Trace;
-
-import me.leolin.shortcutbadger.ShortcutBadger;
 
 import org.apache.cordova.CallbackContext;
 import org.apache.cordova.CordovaPlugin;
-import org.apache.cordova.PluginResult;
 import org.apache.cordova.CordovaWebView;
+import org.apache.cordova.PluginResult;
+import org.apache.cordova.firebase.models.Payload;
+import org.apache.cordova.firebase.notification.NotificationDisplayManager;
+import org.apache.cordova.firebase.utils.SharedPrefsUtils;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
@@ -42,18 +44,12 @@ import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
-
-// Firebase PhoneAuth
 import java.util.concurrent.TimeUnit;
 
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.FirebaseException;
-import com.google.firebase.auth.FirebaseAuthException;
-import com.google.firebase.auth.FirebaseAuthInvalidCredentialsException;
-import com.google.firebase.FirebaseTooManyRequestsException;
-import com.google.firebase.auth.PhoneAuthCredential;
-import com.google.firebase.auth.PhoneAuthProvider;
+import io.fabric.sdk.android.Fabric;
+import me.leolin.shortcutbadger.ShortcutBadger;
+
+// Firebase PhoneAuth
 
 public class FirebasePlugin extends CordovaPlugin {
 
@@ -97,14 +93,14 @@ public class FirebasePlugin extends CordovaPlugin {
     @Override
     public boolean execute(String action, JSONArray args, CallbackContext callbackContext) throws JSONException {
         if (action.equals("initCrashlytics")) {
-          this.initCrashlytics(callbackContext);
-          return true;
+            this.initCrashlytics(callbackContext);
+            return true;
         } else if (action.equals("initAnalytics")) {
-          this.initAnalytics(callbackContext);
-          return true;
+            this.initAnalytics(callbackContext);
+            return true;
         } else if (action.equals("initPerformance")) {
-          this.initPerformance(callbackContext);
-          return true;
+            this.initPerformance(callbackContext);
+            return true;
         } else if (action.equals("getInstanceId")) {
             this.getInstanceId(callbackContext);
             return true;
@@ -256,11 +252,11 @@ public class FirebasePlugin extends CordovaPlugin {
 
         Log.d(TAG, "Initialising Crashlytics");
         try {
-          Fabric.with(context, new Crashlytics());
-          FirebasePlugin.crashlyticsInit = true;
-          callbackContext.success();
-        } catch(Exception e) {
-          callbackContext.error(ERRORINITCRASHLYTICS);
+            Fabric.with(context, new Crashlytics());
+            FirebasePlugin.crashlyticsInit = true;
+            callbackContext.success();
+        } catch (Exception e) {
+            callbackContext.error(ERRORINITCRASHLYTICS);
         }
     }
 
@@ -269,15 +265,15 @@ public class FirebasePlugin extends CordovaPlugin {
 
         Log.d(TAG, "Initialising Analytics");
         try {
-          mFirebaseAnalytics = FirebaseAnalytics.getInstance(context);
-          mFirebaseAnalytics.setAnalyticsCollectionEnabled(true);
-          FirebasePlugin.analyticsInit = true;
-          callbackContext.success();
-        } catch(Exception e) {
-          if(FirebasePlugin.crashlyticsInit()){
-            Crashlytics.logException(e);
-          }
-          callbackContext.error(ERRORINITANALYTICS);
+            mFirebaseAnalytics = FirebaseAnalytics.getInstance(context);
+            mFirebaseAnalytics.setAnalyticsCollectionEnabled(true);
+            FirebasePlugin.analyticsInit = true;
+            callbackContext.success();
+        } catch (Exception e) {
+            if (FirebasePlugin.crashlyticsInit()) {
+                Crashlytics.logException(e);
+            }
+            callbackContext.error(ERRORINITANALYTICS);
         }
     }
 
@@ -286,14 +282,14 @@ public class FirebasePlugin extends CordovaPlugin {
 
         Log.d(TAG, "Initialising Performance");
         try {
-          FirebasePerformance.getInstance().setPerformanceCollectionEnabled(true);
-          FirebasePlugin.performanceInit = true;
-          callbackContext.success();
-        } catch(Exception e) {
-          if(FirebasePlugin.crashlyticsInit()){
-            Crashlytics.logException(e);
-          }
-          callbackContext.error(ERRORINITPERFORMANCE);
+            FirebasePerformance.getInstance().setPerformanceCollectionEnabled(true);
+            FirebasePlugin.performanceInit = true;
+            callbackContext.success();
+        } catch (Exception e) {
+            if (FirebasePlugin.crashlyticsInit()) {
+                Crashlytics.logException(e);
+            }
+            callbackContext.error(ERRORINITPERFORMANCE);
         }
     }
 
@@ -322,8 +318,8 @@ public class FirebasePlugin extends CordovaPlugin {
                         FirebasePlugin.sendToken(currentToken);
                     }
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -349,8 +345,8 @@ public class FirebasePlugin extends CordovaPlugin {
                 try {
                     json.put(key, bundle.get(key));
                 } catch (JSONException e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                     return;
@@ -366,7 +362,7 @@ public class FirebasePlugin extends CordovaPlugin {
     public static void sendNotificationMarkAsRead(Bundle bundle) {
         final CallbackContext callbackContext = FirebasePlugin.notificationMarkAsReadCallbackContext;
 
-        if(callbackContext == null || bundle == null){
+        if (callbackContext == null || bundle == null) {
             return;
         }
 
@@ -376,8 +372,8 @@ public class FirebasePlugin extends CordovaPlugin {
             try {
                 json.put(key, bundle.get(key));
             } catch (JSONException e) {
-                if(FirebasePlugin.crashlyticsInit()){
-                  Crashlytics.logException(e);
+                if (FirebasePlugin.crashlyticsInit()) {
+                    Crashlytics.logException(e);
                 }
                 callbackContext.error(e.getMessage());
                 return;
@@ -457,8 +453,8 @@ public class FirebasePlugin extends CordovaPlugin {
                     String id = FirebaseInstanceId.getInstance().getId();
                     callbackContext.success(id);
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -473,8 +469,8 @@ public class FirebasePlugin extends CordovaPlugin {
                     String token = FirebaseInstanceId.getInstance().getToken();
                     callbackContext.success(token);
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -493,8 +489,8 @@ public class FirebasePlugin extends CordovaPlugin {
                     object.put("isEnabled", areNotificationsEnabled);
                     callbackContext.success(object);
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -507,14 +503,12 @@ public class FirebasePlugin extends CordovaPlugin {
             public void run() {
                 try {
                     Context context = cordova.getActivity();
-                    SharedPreferences.Editor editor = context.getSharedPreferences(KEY, Context.MODE_PRIVATE).edit();
-                    editor.putInt(KEY, number);
-                    editor.apply();
+                    SharedPrefsUtils.putInt(context, KEY, KEY, number);
                     ShortcutBadger.applyCount(context, number);
                     callbackContext.success();
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -526,13 +520,11 @@ public class FirebasePlugin extends CordovaPlugin {
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {
                 try {
-                    Context context = cordova.getActivity();
-                    SharedPreferences settings = context.getSharedPreferences(KEY, Context.MODE_PRIVATE);
-                    int number = settings.getInt(KEY, 0);
+                    int number = SharedPrefsUtils.getInt(cordova.getActivity(), KEY, KEY);
                     callbackContext.success(number);
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -547,8 +539,8 @@ public class FirebasePlugin extends CordovaPlugin {
                     FirebaseMessaging.getInstance().subscribeToTopic(topic);
                     callbackContext.success();
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -563,8 +555,8 @@ public class FirebasePlugin extends CordovaPlugin {
                     FirebaseMessaging.getInstance().unsubscribeFromTopic(topic);
                     callbackContext.success();
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -579,8 +571,8 @@ public class FirebasePlugin extends CordovaPlugin {
                     FirebaseInstanceId.getInstance().deleteInstanceId();
                     callbackContext.success();
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -606,15 +598,15 @@ public class FirebasePlugin extends CordovaPlugin {
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {
                 try {
-                    if(FirebasePlugin.analyticsInit()){
-                      mFirebaseAnalytics.logEvent(name, bundle);
-                      callbackContext.success();
+                    if (FirebasePlugin.analyticsInit()) {
+                        mFirebaseAnalytics.logEvent(name, bundle);
+                        callbackContext.success();
                     } else {
-                      callbackContext.error(ERRORINITANALYTICS);
+                        callbackContext.error(ERRORINITANALYTICS);
                     }
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -626,15 +618,15 @@ public class FirebasePlugin extends CordovaPlugin {
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {
                 try {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(new Exception(message));
-                      callbackContext.success(1);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(new Exception(message));
+                        callbackContext.success(1);
                     } else {
-                      callbackContext.error(ERRORINITCRASHLYTICS);
+                        callbackContext.error(ERRORINITCRASHLYTICS);
                     }
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.log(e.getMessage());
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.log(e.getMessage());
                     }
                     e.printStackTrace();
                     callbackContext.error(e.getMessage());
@@ -648,15 +640,15 @@ public class FirebasePlugin extends CordovaPlugin {
         cordova.getActivity().runOnUiThread(new Runnable() {
             public void run() {
                 try {
-                  if(FirebasePlugin.analyticsInit()){
-                    mFirebaseAnalytics.setCurrentScreen(cordova.getActivity(), name, null);
-                    callbackContext.success();
-                  } else {
-                    callbackContext.error(ERRORINITANALYTICS);
-                  }
+                    if (FirebasePlugin.analyticsInit()) {
+                        mFirebaseAnalytics.setCurrentScreen(cordova.getActivity(), name, null);
+                        callbackContext.success();
+                    } else {
+                        callbackContext.error(ERRORINITANALYTICS);
+                    }
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -668,15 +660,15 @@ public class FirebasePlugin extends CordovaPlugin {
         cordova.getThreadPool().execute(new Runnable() {
             public void run() {
                 try {
-                  if(FirebasePlugin.analyticsInit()){
-                    mFirebaseAnalytics.setUserId(id);
-                    callbackContext.success();
-                  } else {
-                    callbackContext.error(ERRORINITANALYTICS);
-                  }
+                    if (FirebasePlugin.analyticsInit()) {
+                        mFirebaseAnalytics.setUserId(id);
+                        callbackContext.success();
+                    } else {
+                        callbackContext.error(ERRORINITANALYTICS);
+                    }
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -685,23 +677,23 @@ public class FirebasePlugin extends CordovaPlugin {
     }
 
     private void setUserProperty(final CallbackContext callbackContext, final String name, final String value) {
-      if(FirebasePlugin.analyticsInit()){
-        cordova.getThreadPool().execute(new Runnable() {
-            public void run() {
-                try {
-                    mFirebaseAnalytics.setUserProperty(name, value);
-                    callbackContext.success();
-                } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+        if (FirebasePlugin.analyticsInit()) {
+            cordova.getThreadPool().execute(new Runnable() {
+                public void run() {
+                    try {
+                        mFirebaseAnalytics.setUserProperty(name, value);
+                        callbackContext.success();
+                    } catch (Exception e) {
+                        if (FirebasePlugin.crashlyticsInit()) {
+                            Crashlytics.logException(e);
+                        }
+                        callbackContext.error(e.getMessage());
                     }
-                    callbackContext.error(e.getMessage());
                 }
-            }
-        });
-      } else {
-        callbackContext.error(ERRORINITANALYTICS);
-      }
+            });
+        } else {
+            callbackContext.error(ERRORINITANALYTICS);
+        }
     }
 
     private void activateFetched(final CallbackContext callbackContext) {
@@ -711,8 +703,8 @@ public class FirebasePlugin extends CordovaPlugin {
                     final boolean activated = FirebaseRemoteConfig.getInstance().activateFetched();
                     callbackContext.success(String.valueOf(activated));
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -740,15 +732,15 @@ public class FirebasePlugin extends CordovaPlugin {
                     }).addOnFailureListener(new OnFailureListener() {
                         @Override
                         public void onFailure(Exception e) {
-                            if(FirebasePlugin.crashlyticsInit()){
-                              Crashlytics.logException(e);
+                            if (FirebasePlugin.crashlyticsInit()) {
+                                Crashlytics.logException(e);
                             }
                             callbackContext.error(e.getMessage());
                         }
                     });
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -757,41 +749,41 @@ public class FirebasePlugin extends CordovaPlugin {
     }
 
     private void getByteArray(final CallbackContext callbackContext, final String key, final String namespace) {
-      cordova.getThreadPool().execute(new Runnable() {
-        public void run() {
-            try {
-                byte[] bytes = namespace == null ? FirebaseRemoteConfig.getInstance().getByteArray(key)
-                        : FirebaseRemoteConfig.getInstance().getByteArray(key, namespace);
-                JSONObject object = new JSONObject();
-                object.put("base64", Base64.encodeToString(bytes, Base64.DEFAULT));
-                object.put("array", new JSONArray(bytes));
-                callbackContext.success(object);
-            } catch (Exception e) {
-                if(FirebasePlugin.crashlyticsInit()){
-                    Crashlytics.logException(e);
+        cordova.getThreadPool().execute(new Runnable() {
+            public void run() {
+                try {
+                    byte[] bytes = namespace == null ? FirebaseRemoteConfig.getInstance().getByteArray(key)
+                            : FirebaseRemoteConfig.getInstance().getByteArray(key, namespace);
+                    JSONObject object = new JSONObject();
+                    object.put("base64", Base64.encodeToString(bytes, Base64.DEFAULT));
+                    object.put("array", new JSONArray(bytes));
+                    callbackContext.success(object);
+                } catch (Exception e) {
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
+                    }
+                    callbackContext.error(e.getMessage());
                 }
-                callbackContext.error(e.getMessage());
             }
-        }
-      });
+        });
     }
 
     private void getValue(final CallbackContext callbackContext, final String key, final String namespace) {
-      cordova.getThreadPool().execute(new Runnable() {
-        public void run() {
-            try {
-                FirebaseRemoteConfigValue value = namespace == null
-                        ? FirebaseRemoteConfig.getInstance().getValue(key)
-                        : FirebaseRemoteConfig.getInstance().getValue(key, namespace);
-                callbackContext.success(value.asString());
-            } catch (Exception e) {
-                if(FirebasePlugin.crashlyticsInit()){
-                    Crashlytics.logException(e);
+        cordova.getThreadPool().execute(new Runnable() {
+            public void run() {
+                try {
+                    FirebaseRemoteConfigValue value = namespace == null
+                            ? FirebaseRemoteConfig.getInstance().getValue(key)
+                            : FirebaseRemoteConfig.getInstance().getValue(key, namespace);
+                    callbackContext.success(value.asString());
+                } catch (Exception e) {
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
+                    }
+                    callbackContext.error(e.getMessage());
                 }
-                callbackContext.error(e.getMessage());
             }
-        }
-      });
+        });
     }
 
     private void getInfo(final CallbackContext callbackContext) {
@@ -810,7 +802,7 @@ public class FirebasePlugin extends CordovaPlugin {
 
                     callbackContext.success(info);
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
+                    if (FirebasePlugin.crashlyticsInit()) {
                         Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
@@ -829,8 +821,8 @@ public class FirebasePlugin extends CordovaPlugin {
                     FirebaseRemoteConfig.getInstance().setConfigSettings(settings.build());
                     callbackContext.success();
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -848,8 +840,8 @@ public class FirebasePlugin extends CordovaPlugin {
                         FirebaseRemoteConfig.getInstance().setDefaults(defaultsToMap(defaults), namespace);
                     callbackContext.success();
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -912,8 +904,8 @@ public class FirebasePlugin extends CordovaPlugin {
                                 returnResults.put("verificationId", false);
                                 returnResults.put("instantVerification", true);
                             } catch (JSONException e) {
-                                if(FirebasePlugin.crashlyticsInit()){
-                                  Crashlytics.logException(e);
+                                if (FirebasePlugin.crashlyticsInit()) {
+                                    Crashlytics.logException(e);
                                 }
                                 callbackContext.error(e.getMessage());
                                 return;
@@ -940,8 +932,8 @@ public class FirebasePlugin extends CordovaPlugin {
                                 errorMsg = "The SMS quota for the project has been exceeded";
                             }
 
-                            if(FirebasePlugin.crashlyticsInit()){
-                              Crashlytics.logException(e);
+                            if (FirebasePlugin.crashlyticsInit()) {
+                                Crashlytics.logException(e);
                             }
                             callbackContext.error(errorMsg);
                         }
@@ -958,8 +950,8 @@ public class FirebasePlugin extends CordovaPlugin {
                                 returnResults.put("verificationId", verificationId);
                                 returnResults.put("instantVerification", false);
                             } catch (JSONException e) {
-                                if(FirebasePlugin.crashlyticsInit()){
-                                  Crashlytics.logException(e);
+                                if (FirebasePlugin.crashlyticsInit()) {
+                                    Crashlytics.logException(e);
                                 }
                                 callbackContext.error(e.getMessage());
                                 return;
@@ -976,8 +968,8 @@ public class FirebasePlugin extends CordovaPlugin {
                             cordova.getActivity(), // Activity (for callback binding)
                             mCallbacks); // OnVerificationStateChangedCallbacks
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
-                      Crashlytics.logException(e);
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.logException(e);
                     }
                     callbackContext.error(e.getMessage());
                 }
@@ -993,143 +985,143 @@ public class FirebasePlugin extends CordovaPlugin {
 
     private void startTrace(final CallbackContext callbackContext, final String name) {
         final FirebasePlugin self = this;
-        if(FirebasePlugin.performanceInit()){
-          cordova.getThreadPool().execute(new Runnable() {
-              public void run() {
-                  try {
+        if (FirebasePlugin.performanceInit()) {
+            cordova.getThreadPool().execute(new Runnable() {
+                public void run() {
+                    try {
 
-                      Trace myTrace = null;
-                      if (self.traces.containsKey(name)) {
-                          myTrace = self.traces.get(name);
-                      }
+                        Trace myTrace = null;
+                        if (self.traces.containsKey(name)) {
+                            myTrace = self.traces.get(name);
+                        }
 
-                      if (myTrace == null) {
-                          myTrace = FirebasePerformance.getInstance().newTrace(name);
-                          myTrace.start();
-                          self.traces.put(name, myTrace);
-                      }
+                        if (myTrace == null) {
+                            myTrace = FirebasePerformance.getInstance().newTrace(name);
+                            myTrace.start();
+                            self.traces.put(name, myTrace);
+                        }
 
-                      callbackContext.success();
-                  } catch (Exception e) {
-                      if(FirebasePlugin.crashlyticsInit()){
-                        Crashlytics.logException(e);
-                      }
-                      e.printStackTrace();
-                      callbackContext.error(e.getMessage());
-                  }
-              }
-          });
+                        callbackContext.success();
+                    } catch (Exception e) {
+                        if (FirebasePlugin.crashlyticsInit()) {
+                            Crashlytics.logException(e);
+                        }
+                        e.printStackTrace();
+                        callbackContext.error(e.getMessage());
+                    }
+                }
+            });
         } else {
-          callbackContext.error(ERRORINITPERFORMANCE);
+            callbackContext.error(ERRORINITPERFORMANCE);
         }
     }
 
     private void incrementCounter(final CallbackContext callbackContext, final String name, final String counterNamed) {
         final FirebasePlugin self = this;
-        if(FirebasePlugin.performanceInit()){
-          cordova.getThreadPool().execute(new Runnable() {
-              public void run() {
-                  try {
+        if (FirebasePlugin.performanceInit()) {
+            cordova.getThreadPool().execute(new Runnable() {
+                public void run() {
+                    try {
 
-                      Trace myTrace = null;
-                      if (self.traces.containsKey(name)) {
-                          myTrace = self.traces.get(name);
-                      }
+                        Trace myTrace = null;
+                        if (self.traces.containsKey(name)) {
+                            myTrace = self.traces.get(name);
+                        }
 
-                      if (myTrace != null && myTrace instanceof Trace) {
-                          myTrace.incrementCounter(counterNamed);
-                          callbackContext.success();
-                      } else {
-                          callbackContext.error("Trace not found");
-                      }
-                  } catch (Exception e) {
-                      if(FirebasePlugin.crashlyticsInit()){
-                        Crashlytics.logException(e);
-                      }
-                      e.printStackTrace();
-                      callbackContext.error(e.getMessage());
-                  }
-              }
-          });
+                        if (myTrace != null && myTrace instanceof Trace) {
+                            myTrace.incrementCounter(counterNamed);
+                            callbackContext.success();
+                        } else {
+                            callbackContext.error("Trace not found");
+                        }
+                    } catch (Exception e) {
+                        if (FirebasePlugin.crashlyticsInit()) {
+                            Crashlytics.logException(e);
+                        }
+                        e.printStackTrace();
+                        callbackContext.error(e.getMessage());
+                    }
+                }
+            });
         } else {
-          callbackContext.error(ERRORINITPERFORMANCE);
+            callbackContext.error(ERRORINITPERFORMANCE);
         }
     }
 
     private void stopTrace(final CallbackContext callbackContext, final String name) {
         final FirebasePlugin self = this;
-        if(FirebasePlugin.performanceInit()){
-          cordova.getThreadPool().execute(new Runnable() {
-              public void run() {
-                  try {
+        if (FirebasePlugin.performanceInit()) {
+            cordova.getThreadPool().execute(new Runnable() {
+                public void run() {
+                    try {
 
-                      Trace myTrace = null;
-                      if (self.traces.containsKey(name)) {
-                          myTrace = self.traces.get(name);
-                      }
+                        Trace myTrace = null;
+                        if (self.traces.containsKey(name)) {
+                            myTrace = self.traces.get(name);
+                        }
 
-                      if (myTrace != null && myTrace instanceof Trace) { //
-                          myTrace.stop();
-                          self.traces.remove(name);
-                          callbackContext.success();
-                      } else {
-                          callbackContext.error("Trace not found");
-                      }
-                  } catch (Exception e) {
-                      if(FirebasePlugin.crashlyticsInit()){
-                        Crashlytics.logException(e);
-                      }
-                      e.printStackTrace();
-                      callbackContext.error(e.getMessage());
-                  }
-              }
-          });
+                        if (myTrace != null && myTrace instanceof Trace) { //
+                            myTrace.stop();
+                            self.traces.remove(name);
+                            callbackContext.success();
+                        } else {
+                            callbackContext.error("Trace not found");
+                        }
+                    } catch (Exception e) {
+                        if (FirebasePlugin.crashlyticsInit()) {
+                            Crashlytics.logException(e);
+                        }
+                        e.printStackTrace();
+                        callbackContext.error(e.getMessage());
+                    }
+                }
+            });
         } else {
-          callbackContext.error(ERRORINITPERFORMANCE);
+            callbackContext.error(ERRORINITPERFORMANCE);
         }
     }
 
     private void setAnalyticsCollectionEnabled(final CallbackContext callbackContext, final boolean enabled) {
         final FirebasePlugin self = this;
-        if(FirebasePlugin.analyticsInit()){
-          cordova.getThreadPool().execute(new Runnable() {
-              public void run() {
-                  try {
-                      mFirebaseAnalytics.setAnalyticsCollectionEnabled(enabled);
-                      callbackContext.success();
-                  } catch (Exception e) {
-                      if(FirebasePlugin.crashlyticsInit()){
-                        Crashlytics.log(e.getMessage());
-                      }
-                      e.printStackTrace();
-                      callbackContext.error(e.getMessage());
-                  }
-              }
-          });
+        if (FirebasePlugin.analyticsInit()) {
+            cordova.getThreadPool().execute(new Runnable() {
+                public void run() {
+                    try {
+                        mFirebaseAnalytics.setAnalyticsCollectionEnabled(enabled);
+                        callbackContext.success();
+                    } catch (Exception e) {
+                        if (FirebasePlugin.crashlyticsInit()) {
+                            Crashlytics.log(e.getMessage());
+                        }
+                        e.printStackTrace();
+                        callbackContext.error(e.getMessage());
+                    }
+                }
+            });
         } else {
-          callbackContext.error(ERRORINITANALYTICS);
+            callbackContext.error(ERRORINITANALYTICS);
         }
     }
 
     private void setPerformanceCollectionEnabled(final CallbackContext callbackContext, final boolean enabled) {
         final FirebasePlugin self = this;
-        if(FirebasePlugin.performanceInit()){
-          cordova.getThreadPool().execute(new Runnable() {
-              public void run() {
-                  try {
-                      FirebasePerformance.getInstance().setPerformanceCollectionEnabled(enabled);
-                      callbackContext.success();
-                  } catch (Exception e) {
-                      if(FirebasePlugin.crashlyticsInit()){
-                        Crashlytics.log(e.getMessage());
-                      }
-                      e.printStackTrace();
-                      callbackContext.error(e.getMessage());
-                  }
-              }
-          });
+        if (FirebasePlugin.performanceInit()) {
+            cordova.getThreadPool().execute(new Runnable() {
+                public void run() {
+                    try {
+                        FirebasePerformance.getInstance().setPerformanceCollectionEnabled(enabled);
+                        callbackContext.success();
+                    } catch (Exception e) {
+                        if (FirebasePlugin.crashlyticsInit()) {
+                            Crashlytics.log(e.getMessage());
+                        }
+                        e.printStackTrace();
+                        callbackContext.error(e.getMessage());
+                    }
+                }
+            });
         } else {
-          callbackContext.error(ERRORINITPERFORMANCE);
+            callbackContext.error(ERRORINITPERFORMANCE);
         }
     }
 
@@ -1142,9 +1134,9 @@ public class FirebasePlugin extends CordovaPlugin {
                     nm.cancelAll();
                     callbackContext.success();
                 } catch (Exception e) {
-                  if(FirebasePlugin.crashlyticsInit()){
-                    Crashlytics.log(e.getMessage());
-                  }
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.log(e.getMessage());
+                    }
                 }
             }
         });
@@ -1159,7 +1151,7 @@ public class FirebasePlugin extends CordovaPlugin {
                     nm.cancel(id);
                     callbackContext.success();
                 } catch (Exception e) {
-                    if(FirebasePlugin.crashlyticsInit()){
+                    if (FirebasePlugin.crashlyticsInit()) {
                         Crashlytics.log(e.getMessage());
                     }
                 }
@@ -1168,32 +1160,32 @@ public class FirebasePlugin extends CordovaPlugin {
     }
 
     public void scheduleLocalNotification(final CallbackContext callbackContext, final JSONObject params) {
-       cordova.getThreadPool().execute(new Runnable() {
-           public void run() {
-               try {
-                   Context activityContext = cordova.getActivity();
-                   Context appContext = activityContext.getApplicationContext();
+        cordova.getThreadPool().execute(new Runnable() {
+            public void run() {
+                try {
+                    Context activityContext = cordova.getActivity();
+                    Context appContext = activityContext.getApplicationContext();
 
-                   String id = params.getString("id");
-                   String target = params.getString("target");
-                   String username = params.getString("username");
-                   String groupName = params.getString("groupName");
-                   String message = params.getString("message");
-                   String eventType = params.getString("eventType");
-                   String nsound = params.getString("nsound");
-                   String sound = params.getString("sound");
-                   String lights = params.getString("lights");
+                    String id = params.getString("id");
+                    String target = params.getString("target");
+                    String username = params.getString("username");
+                    String groupName = params.getString("groupName");
+                    String message = params.getString("message");
+                    String eventType = params.getString("eventType");
+                    String nsound = params.getString("nsound");
+                    String sound = params.getString("sound");
+                    String lights = params.getString("lights");
 
-                   FirebasePluginMessagingService.displayNotification(activityContext, appContext, id, target, username, groupName, message, eventType, nsound, true, sound, lights);
+                    NotificationDisplayManager.displayNotification(activityContext, appContext, id, target, username, groupName, message, eventType, nsound, true, sound, lights);
 
-                   callbackContext.success();
-               } catch (Exception e) {
-                   if(FirebasePlugin.crashlyticsInit()){
-                       Crashlytics.log(e.getMessage());
-                   }
-                   callbackContext.error(e.getMessage());
-               }
-           }
-       });
-   }
+                    callbackContext.success();
+                } catch (Exception e) {
+                    if (FirebasePlugin.crashlyticsInit()) {
+                        Crashlytics.log(e.getMessage());
+                    }
+                    callbackContext.error(e.getMessage());
+                }
+            }
+        });
+    }
 }
