+++ src/android/OnNotificationOpenReceiver.java
@@ -1,6 +1,5 @@
 package org.apache.cordova.firebase;
 
-import android.app.PendingIntent;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
