+++ src/android/ReplyActivity.java
@@ -8,15 +8,13 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
-
-
 public class ReplyActivity extends Activity implements View.OnClickListener {
     private EditText mMessageTextField;
     private Button mSendButton;
     private Button mCancelButton;
     private String mReplyMessage;
     private static final String VNC_PEER_JID = "vncPeerJid";
-    private static final String  NOTIFY_ID = "id";
+    private static final String NOTIFY_ID = "id";
 
 
     @Override
@@ -36,8 +34,8 @@ public class ReplyActivity extends Activity implements View.OnClickListener {
         if (v == mSendButton) {
             Log.i("VNC", "Send");
             mReplyMessage = mMessageTextField.getText().toString();
-            if(mReplyMessage.equals("")){
-                Toast.makeText(this,"Please enter a message.",Toast.LENGTH_SHORT).show();
+            if (mReplyMessage.equals("")) {
+                Toast.makeText(this, "Please enter a message.", Toast.LENGTH_SHORT).show();
                 return;
             }
             String sender = getIntent().getExtras().getString(VNC_PEER_JID);
