+++ src/android/utils/StringUtils.java
@@ -1,0 +1,14 @@
+package org.apache.cordova.firebase.utils;
+
+import android.content.Context;
+
+public class StringUtils {
+
+    public static String getStringResource(Context activityOrServiceContext, String name) {
+        return activityOrServiceContext.getString(
+                activityOrServiceContext.getResources().getIdentifier(
+                        name, "string", activityOrServiceContext.getPackageName()
+                )
+        );
+    }
+}
